Move UserAuth to util
