+++ src/components/AddPlace/AddPlace.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from "react";
 import * as opencage from "opencage-api-client";
 import { useFirebase } from '../../util/Firebase';
-import { useAuth } from "../Session/UserAuth";
+import { useAuth } from "../../util/UserAuth";
+import { useDebounce } from "../../util/Debounce";
 
 import "./AddPlace.css";
 
@@ -175,27 +176,4 @@ const Row = props => (
   </div>
 );
 
-/**
- * Debounce Hook, returns value after a delay
- * @param {string} value
- * @param {number} delay
- */
-function useDebounce(value, delay) {
-  // State and setters for debounced value
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    // Set debouncedValue to value (passed in) after the specified delay
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => clearTimeout(handler);
-  }, [value]);
-
-  return debouncedValue;
-}
-
 export default AddPlace;
-
-export { useDebounce };
