+++ src/components/Place/MPlaceTable.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { usePlaces } from "../Places/Places";
 import MaterialTable from "material-table";
 import { useFirebase } from '../../util/Firebase';
-import { useAuth } from "../Session/UserAuth";
+import { useAuth } from "../../util/UserAuth";
 
 // Columns used in the table
 const columns = [
